refactor: 이메일, 코드 에러 헨들러 유틸 함수 수정
