+++ src/types/hook.ts
@@ -5,18 +5,6 @@ export interface CompareErrorTypes {
   setPassCompareError: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export interface CodeStatusTypes {
-  email: string;
-  setRegisterEmail: React.Dispatch<React.SetStateAction<string>>;
-  setCodeValidate: React.Dispatch<React.SetStateAction<boolean>>;
-  setAuthDone: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-export interface EmailStatusTypes {
-  setEmailDone: React.Dispatch<React.SetStateAction<boolean>>;
-  setToggle: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
 export interface IntersectionObserverTypes {
   root?: null;
   rootMargin?: string;
